+++ src/Graph/NeighborSelection.cpp
@@ -22,7 +22,7 @@ using namespace std;
 NeighborSelection::NeighborSelection() {
     L1_reg = default_L1_reg;
     L2_reg = default_L2_reg;
-    betaAll = MatrixXd::Ones(1,1);
+    betaAll = MatrixXf::Ones(1,1);
 };
 
 
@@ -37,7 +37,7 @@ NeighborSelection::NeighborSelection(const unordered_map<string, string>& option
     } catch (std::out_of_range& oor) {
         L2_reg = default_L2_reg;
     }
-    betaAll = MatrixXd::Ones(1,1);
+    betaAll = MatrixXf::Ones(1,1);
 };
 
 void NeighborSelection::assertReadyToRun() {
@@ -57,26 +57,26 @@ void NeighborSelection::zeroBetaAt(int idx) {
     beta(idx) = 0;
 }
 
-double NeighborSelection::cost() {
+float NeighborSelection::cost() {
     return 0.5 * (y - X * beta).squaredNorm()/X.rows() + L1_reg * beta.lpNorm<1>() + L2_reg * beta.squaredNorm();
 };
 
-MatrixXd NeighborSelection::derivative() {
+MatrixXf NeighborSelection::derivative() {
     return ((-1.0 * X.transpose() * (y - X * beta)).array() + L1_reg * (beta.array() / beta.cwiseAbs().array()).sum() +
            L2_reg * beta.sum()).matrix();
 };
 
-MatrixXd NeighborSelection::proximal_derivative() {
+MatrixXf NeighborSelection::proximal_derivative() {
     return -1.0 * X.transpose() * (y - X * beta);
 };
 
-MatrixXd NeighborSelection::proximal_operator(VectorXd in, float lr) {
+MatrixXf NeighborSelection::proximal_operator(VectorXf in, float lr) {
     if (L1_reg == 0 && L2_reg == 0){
         return in;
     }
     if (L1_reg != 0 && L2_reg == 0){
-        VectorXd sign = ((in.array()>0).matrix()).cast<double>();//sign
-        sign += -1.0*((in.array()<0).matrix()).cast<double>();
+        VectorXf sign = ((in.array()>0).matrix()).cast<float>();//sign
+        sign += -1.0*((in.array()<0).matrix()).cast<float>();
         in = ((in.array().abs()-lr*L1_reg).max(0)).matrix();//proximal
         return (in.array()*sign.array()).matrix();//proximal multipled back with sign
     }
@@ -84,15 +84,15 @@ MatrixXd NeighborSelection::proximal_operator(VectorXd in, float lr) {
         return in/(1+2*lr*L2_reg);
     }
     else{
-        VectorXd sign = ((in.array()>0).matrix()).cast<double>();
-        sign += -1.0*((in.array()<0).matrix()).cast<double>();
+        VectorXf sign = ((in.array()>0).matrix()).cast<float>();
+        sign += -1.0*((in.array()<0).matrix()).cast<float>();
         in = ((in.array().abs()-lr*L1_reg).max(0)).matrix();
         in = in.array()*sign.array()/(1+2*lr*L2_reg);
         return in.matrix();
     }
 }
 
-void NeighborSelection::updateBetaAll(MatrixXd b) {
+void NeighborSelection::updateBetaAll(MatrixXf b) {
     if (betaAll.rows() == 1){
         betaAll = b;
     }
@@ -102,7 +102,7 @@ void NeighborSelection::updateBetaAll(MatrixXd b) {
     }
 }
 
-MatrixXd NeighborSelection::getBetaAll() {
+MatrixXf NeighborSelection::getBetaAll() {
     return betaAll;
 }
